gcoap proxy: Handle reliable messages
This change enables proxying even when the next hop sends separate
responses.

As bycatch, some pending issues in the proxy code are pointed out. The
Message ID can not be used from the server's message any more (as it
differs in a separate response) and is now shifted if only for better
readable dumps (pending proper creation from the atomic counter). The
token reuse kept as it does greatly simplify things.

After a client's request was ACKed, this now sends the responses in
reliable messages as well.

Currently, clients' requests are only ACKed if the client's CON arrives
a second time (which, given similar timeouts, probably happens all the
time if a package is lost); an advance ACK half-way through the timeout
is still TBD.