proxy: Fix for returning meaningless messages
Before ^, the result of _gcoap_forward_proxy_via_coap was only treated
as an error response (and the server handler returned 0 in non-error
cases). Since it now does occasionally return positive on purpose (when
sending an ACK) and that is passed on, it needs to be careful not to
pass a positive result from an already completed send operation that
would otherwise transmit the not suitably prepared request buffer as a
response.